feat: reorganize notification action definition as to make the list better
